V0.9 'Recipe Ingredients implemented'

+++ Domain/RecipeIngredient.cs
@@ -8,7 +8,7 @@ namespace Domain
         public Recipe Recipe { get; set; }
         public Guid IngredientId { get; set; }
         public Ingredient Ingredient { get; set; }
-        public int Quantity { get; set; }
+        public double Quantity { get; set; }
         public string Measure { get; set; }
     }
 }
